% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mod.R
\name{add_exch_rxns_to_model}
\alias{add_exch_rxns_to_model}
\title{add exchange reactions for metabolites in indicated reaction to sybil model}
\usage{
add_exch_rxns_to_model(model, rxn_idx)
}
\arguments{
\item{model}{sybil model to update}

\item{rxn_idx}{integer indicating reaction whose metabolites should have exchange reactions}
}
\value{
sybil model
}
\description{
add exchange reactions for metabolites in indicated reaction to sybil model
}
\examples{

}
