% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tools.R
\name{check_sets_for_containing}
\alias{check_sets_for_containing}
\title{check to see if set list contains elements of interest}
\usage{
check_sets_for_containing(rxns, sets)
}
\arguments{
\item{rxns}{list if elements to check for}

\item{sets}{list of vectors to search through}
}
\value{
boolean value indicating whether or not all elementst are present
}
\description{
check to see if set list contains elements of interest
}
\examples{

}
