% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falcon_tools.R
\name{clean_rxn_names_in_set}
\alias{clean_rxn_names_in_set}
\title{process names of coupled reactions to replace the names of enzymatic activity reactions to just indicate the genes}
\usage{
clean_rxn_names_in_set(set_list)
}
\arguments{
\item{set_list}{list of lists containing sets of coupled reactions}
}
\value{
processed list of lists
}
\description{
process names of coupled reactions to replace the names of enzymatic activity reactions to just indicate the genes
}
