% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{combine_tables}
\alias{combine_tables}
\title{combine two tables}
\usage{
combine_tables(table_1, table_2, iter_col_1 = 1, iter_col_2 = 1)
}
\arguments{
\item{table_1}{matrix}

\item{table_2}{matrix}

\item{iter_col_1}{which column of table_1 to merge}

\item{iter_col_2}{which column of table_2 to merge}
}
\value{
two-column matrix
}
\description{
combine two tables
}
