% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tools.R
\name{compare_sets}
\alias{compare_sets}
\title{return T/F based on whether or not two set-lists are equivalent}
\usage{
compare_sets(set_1, set_2)
}
\arguments{
\item{set_1}{first set}

\item{set_2}{second set}
}
\value{
boolean indicating whether or not sets are identical
}
\description{
return T/F based on whether or not two set-lists are equivalent
}
\examples{

}
