% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gurobi_tools.R
\name{convert_sybil_to_gurobi}
\alias{convert_sybil_to_gurobi}
\title{description}
\usage{
convert_sybil_to_gurobi(sybil, output = "temp.lp")
}
\arguments{
\item{model}{filename of RData file for sybil model; defaults to 'mutans_model.RData'}
}
\value{
GRB model
}
\description{
description
}
\examples{

}
