% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grb_tools.R
\name{coupling_matrix_from_coupling_vector_list}
\alias{coupling_matrix_from_coupling_vector_list}
\title{Convert the output of GRB_generate_set_lists_cluster to a coupling matrix}
\usage{
coupling_matrix_from_coupling_vector_list(coupling_list, n_react, vars,
  init_sets)
}
\arguments{
\item{coupling_list}{list of list of integers which each indicate the couplings induced by each suppression in PACT}

\item{n_react}{number of reactions}

\item{vars}{names of the reactions; length should equal n_react}

\item{init_sets}{known sets to indicate coupling for in the matrix}
}
\value{
boolean matrix indicating coupling between reactions
}
\description{
Convert the output of GRB_generate_set_lists_cluster to a coupling matrix
}
\examples{

}
