% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raptor_coupling.R
\name{fill_coupling_matrix_from_sets}
\alias{fill_coupling_matrix_from_sets}
\title{given a boolean coupling matrix, ensure that all reactions in each set are indicated to be coupled to each other based on known sets}
\usage{
fill_coupling_matrix_from_sets(mtx, sets)
}
\arguments{
\item{mtx}{boolean coupling matrix to fill}

\item{sets}{sets to indicate in the matrix}
}
\value{
processed boolean matrix
}
\description{
given a boolean coupling matrix, ensure that all reactions in each set are indicated to be coupled to each other based on known sets
}
\examples{

}
