% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tools.R
\name{find_composing_sets}
\alias{find_composing_sets}
\title{find the sets contained by 'rxns'}
\usage{
find_composing_sets(rxns, sets)
}
\arguments{
\item{rxns}{list of elements in a set of interest}

\item{sets}{list of vectors indicating sets}
}
\value{
list of integers indicating the indexes of the sets which are contained in the list of elements
}
\description{
find the sets contained by 'rxns'
}
\examples{

}
