% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gurobi_tools.R
\name{flux_coupling_raptor}
\alias{flux_coupling_raptor}
\title{determine flux coupling between reactions in a Gurobi model}
\usage{
flux_coupling_raptor(model, min_fva_cor = 0.9, fix_frac = 0.1,
  fix_tol_frac = 0.01, bnd_tol = 0.1, stored_obs = 4000,
  cor_iter = 5, cor_check = TRUE, reaction_indexes = c(),
  compare_mtx = FALSE, known_set_mtx = Matrix(data = FALSE, nrow = 1,
  ncol = 1, sparse = TRUE))
}
\arguments{
\item{model}{Gurobi model}

\item{fix_frac}{const value used in fixing flux at non-zero value}

\item{fix_tol_frac}{maximum allowed error when determining whether flux for a reaction is fixed}

\item{bnd_tol}{allowed error in comparing max & min flux}

\item{stored_obs}{maximum number of flux instancecs to store}

\item{cor_iter}{number of iterations after which correlation is considered in checking coupling}

\item{reaction_indexes}{indexes for reactions to check; check all if empty}

\item{compare_mtx}{boolean to indicate whether or not to couple known sets together if any individual reactions are coupled; FALSE by default}

\item{known_set_mtx}{boolean matrix of known coupling between reactions}

\item{min_fva}{minimum correlation value to test for coupling}
}
\value{
boolean matrix indicating whether reactions are coupled to each other (reactions are assigned to rows and columns)
}
\description{
determine flux coupling between reactions in a Gurobi model
}
\examples{

}
