% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_tools.R
\name{gene_set_from_rxn_set}
\alias{gene_set_from_rxn_set}
\title{extrapolate a gene set from a reaction set based on GPR}
\usage{
gene_set_from_rxn_set(GPR, rxn_set_list)
}
\arguments{
\item{GPR}{GPR dataframe}

\item{rxn_set_list}{list of reaction sets}
}
\value{
list of gene sets
}
\description{
extrapolate a gene set from a reaction set based on GPR
}
\examples{

}
