% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falcon_tools.R
\name{generate_falcon_model}
\alias{generate_falcon_model}
\title{Generate FALCON model from base metabolic Sybil model}
\usage{
generate_falcon_model(model, gene_sets = c(), rxn_sets = c())
}
\arguments{
\item{model}{original Sybil model for organism metabolism}

\item{gene_sets}{known gene sets; defaults to empty}

\item{rxn_sets}{known rxn sets; defaults to empty}
}
\value{
Sybil model expanded to include enzymatic activity
}
\description{
Generate FALCON model from base metabolic Sybil model
}
