% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_tools.R
\name{generate_gpr}
\alias{generate_gpr}
\title{generate GPR dataframe}
\usage{
generate_gpr(model)
}
\arguments{
\item{model}{sybil model}
}
\value{
dataframe including association between reactions, genes, and GPR
}
\description{
generate GPR dataframe
}
\examples{

}
