% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tools.R
\name{get_containing_set_idx}
\alias{get_containing_set_idx}
\title{get the position of the set containing all members of a group of elements}
\usage{
get_containing_set_idx(rxns, sets)
}
\arguments{
\item{rxns}{list of elements to search for}

\item{sets}{list of vectors to search through}
}
\value{
integer index of set containing rxns
}
\description{
get the position of the set containing all members of a group of elements
}
\examples{

}
