% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_tools.R
\name{get_genes_from_react_id}
\alias{get_genes_from_react_id}
\title{identify all genes associated with a reaction}
\usage{
get_genes_from_react_id(GPR, rxn_id)
}
\arguments{
\item{GPR}{GPR dataframe}

\item{rxn_id}{string indicating the reaction of interest}
}
\value{
list of strings of genes associated with the reaction of interest
}
\description{
identify all genes associated with a reaction
}
\examples{

}
