% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tools.R
\name{get_list_of_sets_from_mtx}
\alias{get_list_of_sets_from_mtx}
\title{generate sets from coupling mtx}
\usage{
get_list_of_sets_from_mtx(coupling_mtx, init_mtx = NULL)
}
\arguments{
\item{coupling_mtx}{boolean matrix in which TRUE values indicate coupling between reaction or gene on rows and columns}

\item{init_mtx}{boolean matrix of same size and elements as coupling_mtx indicating additional couplings to add; defaults to NULL}
}
\value{
list of sets
}
\description{
generate sets from coupling mtx
}
\examples{

}
