% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mod.R
\name{get_mutans_model}
\alias{get_mutans_model}
\title{generate Streptococcus mutans model}
\usage{
get_mutans_model(reactList = "mutans_model.csv",
  metList = "mutans_model_met.csv")
}
\arguments{
\item{reactList}{filename for list of reactions}

\item{metList}{filename for list of metabolites}
}
\value{
Sybil model
}
\description{
generate Streptococcus mutans model
}
\examples{

}
