% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mod.R
\name{get_PAO_model}
\alias{get_PAO_model}
\title{generate Pseudomonas Aeruginosa model}
\usage{
get_PAO_model(reactList = "PAO1recon1_v23_reactions.csv")
}
\arguments{
\item{reactList}{filename for list of reactions}
}
\value{
Sybil model
}
\description{
generate Pseudomonas Aeruginosa model
}
\examples{

}
