% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tools.R
\name{get_set_idx}
\alias{get_set_idx}
\title{get position of the set containing an element in a list of sets. may return more than one position}
\usage{
get_set_idx(rxn, sets)
}
\arguments{
\item{rxn}{element to search for}

\item{sets}{list of vectors to search through}
}
\value{
list of sets containing rxn
}
\description{
get position of the set containing an element in a list of sets. may return more than one position
}
\examples{

}
