% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mod.R
\name{get_yeast_maranas_model}
\alias{get_yeast_maranas_model}
\title{generate Saccharomyces cerevisiae model}
\usage{
get_yeast_maranas_model(reactList = "S7 Model iSce926.tsv",
  metList = "S7 Model iSce926_met.tsv")
}
\arguments{
\item{reactList}{filename for list of reactions}

\item{metList}{filename for list of metabolites}
}
\value{
Sybil model
}
\description{
generate Saccharomyces cerevisiae model
}
\examples{

}
