% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gurobi_tools.R
\name{GRB_flux_coupling_raptor_wrapper}
\alias{GRB_flux_coupling_raptor_wrapper}
\title{wrapper function for flux coupling}
\usage{
GRB_flux_coupling_raptor_wrapper(i, vars, model_og,
  reaction_indexes = 1:length(vars), compare_mtx = FALSE,
  init_coupling_mtx = c(), file_output = NULL)
}
\arguments{
\item{i}{integer indicating index of the reaction to suppress}

\item{vars}{vector of reaction ids}

\item{model_og}{Gurobi model}

\item{reaction_indexes}{integers indicating reactions to run flux coupling over; defaults to 1:length(vars)}

\item{compare_mtx}{boolean to indicate whether or not to couple known sets together if any individual reactions are coupled; FALSE by default}

\item{init_coupling_mtx}{boolean matrix of known coupling between reactions; empty by default}

\item{file_output}{filename to output coupling vector to}
}
\value{
list of integers which indicate positions of TRUE values in couling matrix
}
\description{
wrapper function for flux coupling
}
\examples{

}
