% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gurobi_tools.R
\name{GRB_generate_falcon_model}
\alias{GRB_generate_falcon_model}
\title{Generate Gurobi falcon model. First generates model in sybil then converts the model to Gurobi in order to add additional
constraints on flux}
\usage{
GRB_generate_falcon_model(sybil_model, falcon_model = FALSE,
  r0_gene_set = c(), r0_rxn_set_list = c())
}
\arguments{
\item{sybil_model}{base model to build falcon model from}

\item{falcon_model}{boolean value indicating whether or not the base model has enzyme activity already. If TRUE, then the function will simply add contstraints to the falcon model}

\item{r0_gene_set}{list of genes passed in to the falcon-model building method (generate_falcon_model)}

\item{r0_rxn_set_list}{list of reaction sets passed in to the falcon-model building method (generate_falcon_model)}
}
\value{
Gurobi model based on metabolic model with enzymatic acctivity
}
\description{
Generate Gurobi falcon model. First generates model in sybil then converts the model to Gurobi in order to add additional
constraints on flux
}
\examples{

}
