% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gurobi_tools.R
\name{GRB_get_rxn_idx}
\alias{GRB_get_rxn_idx}
\title{get the position of a reaction along its axis on the S matrix}
\usage{
GRB_get_rxn_idx(model, rxn)
}
\arguments{
\item{model}{Gurobi model}

\item{rxn}{reaction id to get the index of}
}
\value{
integer indicating the index of the reaction along the S matrix
}
\description{
get the position of a reaction along its axis on the S matrix
}
\examples{

}
