% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gurobi_tools.R
\name{GRB_maximize}
\alias{GRB_maximize}
\title{flux-balance optimization function for Gurobi}
\usage{
GRB_maximize(model_og, obj, suppress = c(), max = TRUE)
}
\arguments{
\item{model_og}{Gurobi model}

\item{obj}{integer indicating the index of the objective reaction}

\item{suppress}{list of integers indicating which reactions to suppress before optimizing}

\item{max}{boolean indicating whether to maximize or minimize the objective. TRUE indicates maximization, FALSE indicates minimization. DEFAULT: TRUE}
}
\description{
flux-balance optimization function for Gurobi
}
\examples{

}
