% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gurobi_tools.R
\name{GRB_mutans_model}
\alias{GRB_mutans_model}
\title{Gurobi S. mutans model}
\usage{
GRB_mutans_model(model = "mutans_model.RData")
}
\arguments{
\item{model}{filename of RData file for sybil model; defaults to 'mutans_model.RData'}
}
\value{
GRB model
}
\description{
Gurobi S. mutans model
}
\examples{

}
