% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gurobi_tools.R
\name{GRB_pao_model}
\alias{GRB_pao_model}
\title{Gurobi P. aeruginosa model}
\usage{
GRB_pao_model(model = "pao_model.RData")
}
\arguments{
\item{model}{filename of RData file for sybil model; defaults to 'pao_model.RData'}
}
\value{
GRB model
}
\description{
Gurobi P. aeruginosa model
}
\examples{

}
