% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gurobi_tools.R
\name{GRB_yeast_model}
\alias{GRB_yeast_model}
\title{Gurobi S. cerevisiae model}
\usage{
GRB_yeast_model(model = "yeast_model.RData")
}
\arguments{
\item{model}{filename of RData file for sybil model; defaults to 'yeast_model.RData'}
}
\value{
GRB model
}
\description{
Gurobi S. cerevisiae model
}
\examples{

}
