% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grb_tools.R
\name{identify_intermediate_uncoupled}
\alias{identify_intermediate_uncoupled}
\title{identify which G sets and R sets within the same G* or R* set never couple to each other}
\usage{
identify_intermediate_uncoupled(full_coupling_mtx, coupling_mtx, n_react)
}
\arguments{
\item{full_coupling_mtx}{boolean matrix indicating fully coupled G* or R* sets}

\item{coupling_mtx}{boolean matrix indicating only observed couplings}

\item{n_react}{number of reactions}
}
\value{
boolean matrix with TRUE values indicating which reactions in the same G* or R* set never couple
}
\description{
identify which G sets and R sets within the same G* or R* set never couple to each other
}
\examples{

}
