% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falcon_tools.R
\name{isolate_gene_matrix}
\alias{isolate_gene_matrix}
\title{reduce coupling matrix to only contain the reactions for enzymatic activity}
\usage{
isolate_gene_matrix(coupling_matrix)
}
\arguments{
\item{coupling_matrix}{Matrix of TRUE/FALSE values}
}
\value{
Matrix of TRUE/FALSE values
}
\description{
reduce coupling matrix to only contain the reactions for enzymatic activity
}
