% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falcon_tools.R
\name{isolate_rxn_matrix}
\alias{isolate_rxn_matrix}
\title{reduce coupling matrix to only the metabolic reactions}
\usage{
isolate_rxn_matrix(coupling_matrix)
}
\arguments{
\item{coupling_matrix}{Matrix of TRUE/FALSE values}
}
\value{
Matrix of TRUE/FALSE values
}
\description{
reduce coupling matrix to only the metabolic reactions
}
