% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raptor_coupling.R
\name{optimize_rxn}
\alias{optimize_rxn}
\title{minimize or maximize flux through a reaction in a given Gurobi model}
\usage{
optimize_rxn(model, rxn, max)
}
\arguments{
\item{model}{Gurobi model}

\item{rxn}{string indicating reaction to optimize flux for}

\item{max}{boolean value indicating whether or not to maximize flux. FALSE indicates minimization}
}
\value{
fluxes for all reactions in models
}
\description{
minimize or maximize flux through a reaction in a given Gurobi model
}
\examples{

}
