% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{read_coupling_csv}
\alias{read_coupling_csv}
\title{Read in coupling file}
\usage{
read_coupling_csv(filename, coupling_vector = c(),
  completed_idxs = c())
}
\arguments{
\item{filename}{name of file which PACT output was written to}

\item{coupling_vector}{optional coupling vector, if warm-starting}

\item{completed_idxs}{optional copmpleted indexes, if warm starting}
}
\value{
list of coupling vectors (list of lists containing positions of TRUE values in coupling matrix) and completed indexes (list of integers of suppressions already read)
}
\description{
Read in coupling file
}
