% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raptor_coupling.R
\name{set_model_bounds}
\alias{set_model_bounds}
\title{change the bounds on a Gurobi model}
\usage{
set_model_bounds(model, rxn, UB, LB)
}
\arguments{
\item{model}{Gurobi model}

\item{rxn}{string indicating the reaction whose bounds should be altered}

\item{UB}{new upper bound for reaction}

\item{LB}{new lower bound for reaction}
}
\value{
new model with altered bounds
}
\description{
change the bounds on a Gurobi model
}
\examples{

}
