% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gurobi_tools.R
\name{solve_model}
\alias{solve_model}
\title{Solve Gurobi model}
\usage{
solve_model(model, obj_idx, sense = "max")
}
\arguments{
\item{model}{Gurobi model}
}
\value{
solution to optimization problem
}
\description{
Solve Gurobi model
}
\examples{

}
